refactor: extract CliUninstallHandler from App.xaml.cs
The --uninstall CLI path (RunCliUninstallAsync, CliRedact, AttachConsole
P/Invoke) had no dependency on App instance state. Moving it to a
dedicated static class removes ~150 lines from App and gives the
headless uninstall entry point a clear, named home.
